feat(code-server): add python venv for gitla formation (#245)
venv is needed in order to be able to install the dependencies.
With the code added, it is automatically installed and configured so that the participants
do not have to do any manipulation regarding python and venv.

Meanwhile the pods are restarted every day, which means that the venv folder will be lost and every day participants will have to reinstall the dependencies if necessary.
Righ now it is ok, but a better alternative would be to create a persistent volume.