revert(gitla): remove default venv creation
Default venv creation for gitla formation is not relevent, as it was installed in default coder folder and every trainee has a personal home folder created, so it was not mgically working.

So the default venv creation has been removed from the docker file, and the TPs has been updated to create it manually (easier that way).